refactor: :recycle: move `cdp` constant into `constants.py`
Within the `check_datapackage` module.
