+++ src/seedcase_sprout/core/check_datapackage/__init__.py
@@ -5,13 +5,13 @@ from .check_error_matcher import CheckErrorMatcher
 from .check_package_properties import check_package_properties
 from .check_properties import check_properties
 from .check_resource_properties import check_resource_properties
-from .exclude_matching_errors import exclude_matching_errors
-from .required_fields import (
+from .constants import (
     PACKAGE_RECOMMENDED_FIELDS,
     PACKAGE_REQUIRED_FIELDS,
     RESOURCE_REQUIRED_FIELDS,
     RequiredFieldType,
 )
+from .exclude_matching_errors import exclude_matching_errors
 
 __all__ = [
     "CheckError",
