+++ src/seedcase_sprout/core/check_datapackage/add_package_recommendations.py
@@ -1,6 +1,7 @@
-from seedcase_sprout.core.check_datapackage.config import NAME_PATTERN, SEMVER_PATTERN
-from seedcase_sprout.core.check_datapackage.required_fields import (
+from seedcase_sprout.core.check_datapackage.constants import (
+    NAME_PATTERN,
     PACKAGE_RECOMMENDED_FIELDS,
+    SEMVER_PATTERN,
 )
