+++ src/seedcase_sprout/core/check_datapackage/add_resource_recommendations.py
@@ -1,4 +1,4 @@
-from seedcase_sprout.core.check_datapackage.config import NAME_PATTERN
+from seedcase_sprout.core.check_datapackage.constants import NAME_PATTERN
 
 
 def add_resource_recommendations(schema: dict) -> dict:
