+++ src/seedcase_sprout/core/check_datapackage/constants.py
@@ -1,11 +1,31 @@
 from importlib.resources import files
 from pathlib import Path
 
+from seedcase_sprout.core.check_datapackage.required_fields import RequiredFieldType
+
+COMPLEX_VALIDATORS = {"allOf", "anyOf", "oneOf"}
+
 DATA_PACKAGE_SCHEMA_PATH: Path = files(
     "seedcase_sprout.core.check_datapackage.schemas"
 ).joinpath("data-package-schema.json")
 
 NAME_PATTERN = r"^[a-z0-9._-]+$"
+
+PACKAGE_RECOMMENDED_FIELDS = {
+    "name": RequiredFieldType.str,
+    "id": RequiredFieldType.str,
+    "licenses": RequiredFieldType.list,
+}
+
+PACKAGE_REQUIRED_FIELDS = {
+    "resources": RequiredFieldType.list,
+}
+
+RESOURCE_REQUIRED_FIELDS = {
+    "name": RequiredFieldType.str,
+    "path": RequiredFieldType.str,
+    "data": RequiredFieldType.any,
+}
 
 # From https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string
 SEMVER_PATTERN = (
